ai-audio-microservice/requirements.txt
Flask
Flask-Cors
python-dotenv
gunicorn
torch
transformers
pydub
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# winget install ffmpeg